Moves exploration space API methods to separate file
